requirements.txt
Django==4.2.7
django-ninja==1.0.1
pydantic==2.5.0
PyJWT==2.8.0
django-redis==5.4.0
redis==5.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-debug-toolbar==4.2.0
Pillow==10.1.0
